websockets

+++ controllers/postsController.js
@@ -18,7 +18,7 @@ class postsController {
 
   async addPost(req, res) {
     try {
-      const user = await User.findOne({ id: req.user.id });
+      const user = await User.findOne({ _id: req.user.id });
 
       if (!user) return res.status(404).json({ message: "User not found" });
 
@@ -43,7 +43,7 @@ class postsController {
       });
 
       await post.save();
-      return res.json({ message: "The post has been created" });
+      return res.json({ message: "Post created" });
     } catch (error) {
       console.log(error);
       return res.status(500).json({ message: "Post upload error" });
@@ -52,7 +52,7 @@ class postsController {
 
   async deletePost(req, res) {
     try {
-      const user = await User.findOne({ id: req.user.id });
+      const user = await User.findOne({ _id: req.user.id });
       user.posts -= 1;
       await user.save();
 
@@ -65,19 +65,21 @@ class postsController {
 
       const comments = await Comment.find({ postId: post._id });
 
-      comments.forEach(async (item) => {
-        const comment = await Comment.findOne({ _id: item._id });
+      if (comments) {
+        comments.forEach(async (item) => {
+          const comment = await Comment.findOne({ _id: item._id });
 
-        if (comment?.image) {
-          const path = req.filePath + "/" + comment.image;
-          await fs.unlink(path);
-        }
-      });
+          if (comment?.image) {
+            const path = req.filePath + "/" + comment.image;
+            await fs.unlink(path);
+          }
+        });
 
-      await Comment.deleteMany({ postId: post._id });
+        await Comment.deleteMany({ postId: post._id });
+      }
 
       await post.remove();
-      return res.json({ message: "The post has been deleted" });
+      return res.json({ message: "Post deleted" });
     } catch (error) {
       console.log(error);
       return res.status(500).json({ message: "Post deletion error" });
@@ -93,7 +95,7 @@ class postsController {
       post.liked = liked;
       await post.save();
 
-      return res.json({ message: "The post has been updated" });
+      return res.json({ message: "Post updated" });
     } catch (error) {
       console.log(error);
       return res.status(500).json({ message: "Post updated error" });
